Multi-expression Concept
I'm still struggling with the best way to represent this, but this
isn't bad. Basically, there are times where we require a call to be
explicit and other times where it's implicit, but it all has to do
with /immediate/ context (i.e. a preceding `if` token).

For example, an if condition is parsed as a single expression, NOT a
callee unless explicitly written (within parens). This only affects
the collection of arguments. So, we could also distinguish this
through expressions that may collect arguments and form a call vs
"other" expressions; this is the idea of the multi-expression---a
multi expression may collect arguments, but a regular expression may
not.